# openapi/delta-backend.openapi.yaml
openapi: 3.0.0
paths:
...
  /api/business/catalog/all:
    get:
      operationId: BusinessCatalog_getCatalogAll
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: object
      security:
        - businessSession: []
      summary: Get catalog data for business administration
      tags:
        - Business Catalog
...
  /api/business/catalog/documents/access:
    post:
      operationId: BusinessCatalog_createDocumentAccess
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentAccessDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                type: object
      security:
        - businessSession: []
          csrfToken: []
      summary: Create an administrative signed document URL
      tags:
        - Business Catalog
...
  /api/catalog/library:
    post:
      operationId: Catalog_getCatalogLibrary
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CatalogLibraryDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                type: object
      summary: Get selected company catalog libraries
      tags:
        - Catalog
  /api/catalog/access:
    post:
      operationId: Catalog_createAccess
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CatalogAccessDto"
      responses:
        "201":
          description: ""
      summary: Record a catalog access inquiry
      tags:
        - Catalog
  /api/catalog/access/google:
    post:
      operationId: Catalog_createGoogleAccess
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CatalogGoogleAccessDto"
      responses:
        "201":
          description: ""
      summary: Create catalog access with Google
      tags:
        - Catalog
  /api/catalog/access/google/redirect:
    post:
      operationId: Catalog_createGoogleRedirectAccess
      parameters: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/CatalogGoogleRedirectDto"
      responses:
        "303":
          description: Redirect to the frontend home page.
      summary: Handle Google catalog access redirect
      tags:
        - Catalog
  /api/catalog/access/me:
    get:
      operationId: Catalog_getAccessMe
      parameters: []
      responses:
        "200":
          description: ""
      security:
        - catalogAccess: []
      summary: Get the current catalog access session
      tags:
        - Catalog
  /api/catalog/access/request-otp:
    post:
      operationId: Catalog_requestAccessOtp
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CatalogOtpRequestDto"
      responses:
        "201":
          description: ""
      summary: Request a catalog access OTP
      tags:
        - Catalog
  /api/catalog/access/verify-otp:
    post:
      operationId: Catalog_verifyAccessOtp
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CatalogVerifyOtpDto"
      responses:
        "201":
          description: ""
      summary: Verify a catalog access OTP
      tags:
        - Catalog
  /api/catalog/documents/access:
    post:
      operationId: Catalog_createDocumentAccess
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentAccessDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                type: object
      summary: Create a signed catalog document URL
      tags:
        - Catalog
  /api/catalog/access/logout:
    post:
      operationId: Catalog_logoutAccess
      parameters: []
      responses:
        "201":
          description: ""
      security:
        - catalogAccess: []
      summary: Log out of catalog access
      tags:
        - Catalog
...
info:
  title: Delta Backend API
  description: Public catalog and product APIs, catalog access flows,
    authenticated business administration, and internal administration
    endpoints.
  version: 0.0.1
  contact: {}
...
servers: []
components:
  securitySchemes:
    businessSession:
      type: apiKey
      in: cookie
      name: business_session
      description: Business session cookie. Production uses __Host-business_session.
    catalogAccess:
      type: apiKey
      in: cookie
      name: catalog_access
      description: Catalog access session cookie.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: CSRF token returned by business authentication.
    internalAdmin:
      type: apiKey
      in: header
      name: X-Backend-Admin-Token
      description: Token for the optional internal administration API.
  schemas:
...
    DocumentAccessDto:
      type: object
      properties:
        company_slug:
          type: string
        category_slug:
          type: string
        document_slug:
          type: string
        action:
          enum:
            - preview
            - download
          type: string
      required:
        - company_slug
        - document_slug
        - action
...
    CatalogLibraryDto:
      type: object
      properties:
        company_slugs:
          minItems: 1
          type: array
          items:
            type: string
      required:
        - company_slugs
    CatalogAccessDto:
      type: object
      properties:
        name:
          type: string
          maxLength: 120
        mobile:
          type: string
          maxLength: 30
        email:
          type: string
          maxLength: 160
          format: email
        message:
          type: string
          maxLength: 1000
      required:
        - name
        - mobile
    CatalogGoogleAccessDto:
      type: object
      properties:
        id_token:
          type: string
          maxLength: 4096
      required:
        - id_token
    CatalogGoogleRedirectDto:
      type: object
      properties:
        credential:
          type: string
          maxLength: 4096
        g_csrf_token:
          type: string
          maxLength: 512
      required:
        - credential
        - g_csrf_token
    CatalogOtpRequestDto:
      type: object
      properties:
        name:
          type: string
          maxLength: 120
        mobile:
          type: string
          maxLength: 30
        email:
          type: string
          maxLength: 160
          format: email
        message:
          type: string
          maxLength: 1000
        channel:
          enum:
            - whatsapp
            - email
          type: string
      required:
        - name
        - channel
    CatalogVerifyOtpDto:
      type: object
      properties:
        challenge_id:
          type: string
          maxLength: 40
        mobile:
          type: string
          maxLength: 30
        email:
          type: string
          maxLength: 160
          format: email
        otp:
          type: string
          maxLength: 12
      required:
        - challenge_id
        - otp
...